pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automatic-bulk-email-sender"
version = "0.1.0"
description = "Automatic bulk email sender with Jinja2 templates, Gmail/SendGrid support"
requires-python = ">=3.8"
dependencies = [
    "Jinja2>=3.0.0",
    "email-validator>=1.3.0",
    "google-auth-oauthlib>=0.7.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.50.0",
    "sendgrid>=6.9.0",
    "python-dotenv>=0.19.0",
    "click>=8.0.0",
    "pydantic>=1.9.0",
    "requests>=2.28.0",
    "PyYAML>=5.4.0",
]

[project.scripts]
email-sender = "email_sender.cli:main"
